backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohappyeyeballs==2.4.6
    # via aiohttp
aiohttp==3.11.12
    # via litellm
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.45.2
    # via -r requirements/requirements.in
anyio==4.8.0
    # via
    #   anthropic
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
boto3==1.36.16
    # via -r requirements/requirements.in
botocore==1.36.16
    # via
    #   boto3
    #   s3transfer
brotli==1.1.0
    # via whitenoise
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   pdfminer-six
    #   requests
click==8.1.8
    # via
    #   -r requirements/requirements.in
    #   litellm
cryptography==44.0.0
    # via
    #   pdfminer-six
    #   pyjwt
distro==1.9.0
    # via
    #   anthropic
    #   openai
dj-rest-auth==7.0.1
    # via -r requirements/requirements.in
dj-stripe==2.8.4
    # via -r requirements/requirements.in
django==5.1.6
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-cors-headers
    #   django-otp
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth[mfa,socialaccount]==0.63.3
    # via -r requirements/requirements.in
django-anymail[mailgun]==12.0
    # via -r requirements/requirements.in
django-cors-headers==4.7.0
    # via -r requirements/requirements.in
django-environ==0.12.0
    # via -r requirements/requirements.in
django-otp==1.5.4
    # via -r requirements/requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-api-key==3.0.0
    # via -r requirements/requirements.in
djangorestframework-simplejwt==5.4.0
    # via -r requirements/requirements.in
docutils==0.21.2
    # via -r requirements/requirements.in
drf-spectacular==0.28.0
    # via -r requirements/requirements.in
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.17.0
    # via huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via huggingface-hub
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.24.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.160.0
    # via google-generativeai
google-auth==2.38.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.4
    # via -r requirements/requirements.in
googleapis-common-protos==1.67.0rc1
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.70.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   anthropic
    #   litellm
    #   openai
huggingface-hub==0.28.1
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via litellm
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.5
    # via litellm
jiter==0.8.2
    # via
    #   anthropic
    #   openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via
    #   drf-spectacular
    #   litellm
jsonschema-specifications==2024.10.1
    # via jsonschema
litellm==1.60.8
    # via -r requirements/requirements.in
lxml==5.3.2
    # via
    #   python-docx
    #   python-pptx
markupsafe==3.0.2
    # via jinja2
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.24.3
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.61.1
    # via
    #   -r requirements/requirements.in
    #   litellm
openpyxl==3.1.5
    # via -r requirements/requirements.in
packaging==24.2
    # via
    #   djangorestframework-api-key
    #   huggingface-hub
    #   pytesseract
pandas==2.0.3
    # via -r requirements/requirements.in
pdfminer-six==20250327
    # via pdfplumber
pdfplumber==0.11.6
    # via -r requirements/requirements.in
pillow==10.4.0
    # via
    #   -r requirements/requirements.in
    #   pdfplumber
    #   pytesseract
    #   python-pptx
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.10
    # via -r requirements/requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   anthropic
    #   google-generativeai
    #   litellm
    #   openai
pydantic-core==2.27.2
    # via pydantic
pyjwt[crypto]==2.10.1
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pymupdf==1.25.5
    # via -r requirements/requirements.in
pyparsing==3.2.1
    # via httplib2
pypdfium2==4.30.1
    # via pdfplumber
pytesseract==0.3.13
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-docx==1.1.2
    # via -r requirements/requirements.in
python-dotenv==1.0.1
    # via litellm
python-pptx==1.0.2
    # via -r requirements/requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   huggingface-hub
qrcode==8.0
    # via django-allauth
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   django-allauth
    #   django-anymail
    #   google-api-core
    #   huggingface-hub
    #   requests-oauthlib
    #   stripe
    #   tiktoken
requests-oauthlib==2.0.0
    # via django-allauth
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
s3transfer==0.11.2
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   openai
sqlparse==0.5.3
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
tiktoken==0.8.0
    # via litellm
tokenizers==0.21.0
    # via litellm
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   anthropic
    #   anyio
    #   google-generativeai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   python-pptx
    #   referencing
tzdata==2025.1
    # via pandas
uritemplate==4.1.1
    # via
    #   drf-spectacular
    #   google-api-python-client
urllib3==2.3.0
    # via
    #   botocore
    #   django-anymail
    #   requests
whitenoise[brotli]==6.9.0
    # via -r requirements/requirements.in
xlsxwriter==3.2.2
    # via python-pptx
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
